requirements.txt
aiohttp-socks==0.7.1
aiosignal==1.2.0
alabaster==0.7.13
alembic==1.7.5
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
brotlipy 
certifi==2022.12.7
cffi==1.14.6
charset-normalizer 
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==14.3
comm==0.1.2
contourpy==1.0.7
coverage==7.2.2
cryptography==3.4.7
cycler==0.11.0
databricks-cli==0.17.5
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
docker==6.0.1
docopt==0.6.2
docutils==0.19
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
flake8==6.0.0
Flask==2.2.3
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.2.0
fs==2.4.16
gitdb==4.0.10
GitPython==3.1.31
greenlet==1.1.2
gunicorn==20.1.0
humanfriendly==10.0
idna 
imagesize==1.4.1
imbalanced-learn==0.10.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==2.0.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
llvmlite==0.39.1
Mako==1.1.6
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mergedeep==1.3.4
mistune==2.0.5
mlflow==2.2.2
multidict==5.2.0
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.1.0
packaging 
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
patsy 
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pooch 
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.22.1
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycodestyle==2.10.0
pycparser==2.20
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks 
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-socks==2.0.0
pytz 
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.1
QtPy==2.3.0
querystring-parser==1.2.4
requests 
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.1.0
selenium-wire==2.1.2
Send2Trash==1.8.0
shap==0.41.0
six 
slicer==0.0.7
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.20
sqlparse==0.4.3
stack-data==0.6.2
statsmodels 
tabulate==0.9.0
tenacity==6.3.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions 
tzdata==2023.3
ujson==4.3.0
uri-template==1.2.0
urllib3 
user-agent==0.1.10
verboselogs==1.7
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
win32-setctime==1.0.4
wsproto==1.0.0
zipp==3.14.0
